+++ RELEASENOTES.md
@@ -1,3 +1,9 @@
+## 4.1.0
+
+* use .mulle folder
+* make mulle-test on its own more powerful and convenient
+
+
 ### 4.0.10
 
 * ensure mulle-testallocator is first, dont ignore quickstatus errors anymore